requirements.txt
psycopg2==2.9.3
Flask==2.1.2
flask-restx==0.5.1
python-dotenv==0.20.0
